Refactor `CustomContractPopulator` by extracting method and parameter order resolution logic into separate utility classes
